pyproject.toml
[tool.poetry]
name = "album-manager-cli"
version = "0.1.0"
description = "TBD"
authors = ["Rodrigo Sousa <rod_sousa@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "8.4.0"
loguru = "^0.7.2"
mypy = "^1.13.0"
questionary = "^2.0.1"
opencv-python = "^4.10.0.84"
pillow-heif = "0.3.2"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
mypy = "^1.13.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
